Fix various FindBugs problems and warnings
(cherry picked from commit 3d542c0073af98b13fc6e132a0413907f85a24f8)

Conflicts:
	java/squeek/veganoption/blocks/BlockBasin.java
	java/squeek/veganoption/blocks/renderers/RenderBasin.java
	java/squeek/veganoption/blocks/tiles/TileEntityBasin.java
	java/squeek/veganoption/content/crafting/PistonCraftingHandler.java
	java/squeek/veganoption/content/recipes/InputItemStack.java
	java/squeek/veganoption/content/recipes/PistonCraftingRecipe.java
	java/squeek/veganoption/integration/nei/PistonCraftingHandler.java

+++ java/squeek/veganoption/integration/waila/ProviderRettable.java
@@ -26,7 +26,7 @@ public class ProviderRettable implements IWailaDataProvider
 	public List<String> getWailaBody(ItemStack itemStack, List<String> toolTip, IWailaDataAccessor accessor, IWailaConfigHandler config)
 	{
 		BlockRettable blockRettable = (BlockRettable) accessor.getBlock();
-		float rettingPercent = (float) blockRettable.getRettingPercentFromMeta(accessor.getMetadata());
+		float rettingPercent = blockRettable.getRettingPercentFromMeta(accessor.getMetadata());
 		if (rettingPercent >= 1)
 			toolTip.add(LangHelper.translate("waila.retted"));
 		else
